Broke out larger functions into more simple ones and separated across files.
