+++ src/components/drawCard.jsx
@@ -1,0 +1,4 @@
+export function drawCard(shoe) {
+    const drawnCard = shoe.shift();
+    return drawnCard;
+};
